Everything builds with .NET 5 now and VS Code can be used directly.
