+++ .gitignore
@@ -40,6 +40,8 @@ local.properties
 *.suo
 *.user
 *.sln.docstates
+.vs
+
 
 # Build results
 [Dd]ebug/
