+++ ItemCollage/Options.cs
