+++ ItemCollage/Properties/Settings.settings
